+++ src/App.js
@@ -6,6 +6,7 @@ import ColorPicker from './components/colorPicker';
 import { SQUARES } from './constants/squares'
 import { faceletString } from './utils/cubejsCompatibility';
 import { validateColors } from './utils/validateColors';
+import { moveToKey } from './utils/movesKey';
 
 export const ColorPickerContext = createContext({
   setShowCp: () => undefined,
@@ -24,6 +25,7 @@ function App() {
   const [inputs, setInputs] = useState(SQUARES)
   const [disabled, setDisabled] = useState(false)
   const [solution, setSolution] = useState()
+  const [solving, setSolving] = useState(false)
 
   useEffect(() => {
     colorPicker(showCp, setShowCp, setColorPickerColor)
@@ -31,7 +33,11 @@ function App() {
   }, [showCp, colorPickerCube, colorPickerColor, inputs])
 
   const solve = async () => {
-    await fetch(`http://localhost:5000/solveCube?cube=${faceletString(inputs)}`).then(res => res.json()).then(data => setSolution(data.join(' ')))
+    setSolving(true)
+    await fetch(`http://localhost/api/solveCube?facelets=${faceletString(inputs)}`)
+      .then(res => res.json())
+      .then(data => setSolution(data))
+      .finally(() => setSolving(false))
   }
 
   return (
@@ -52,6 +58,7 @@ function App() {
         }}>
         <Cube />
         <div
+          id='color-picker-wrapper'
           style={{
             alignSelf: 'center',
             marginLeft: 100
@@ -62,15 +69,33 @@ function App() {
             : <button
                 onClick={solve}
                 disabled={disabled}
-              >solve</button>
+              >{solving ? '...solving' : 'solve'}</button>
           }
         </div>
       </div>
       <div
+        className='solution-container'
         style={{
-          padding: 50
+          padding: '50px 50px 0px'
         }}>
-        {solution}
+        {solution && (
+          <>
+            <h4>solution:</h4>
+            <p>{solution.join(' ')}</p>
+          </>
+        )}
+      </div>
+      <div
+        className='solution-container'
+        style={{
+          padding: '0px 50px 50px'
+        }}>
+        {solution && (
+          <>
+            <h4>steps:</h4>
+            {solution.map((move, idx) => <p key={idx}>{idx + 1}. <strong>{move}: </strong>{moveToKey(move)}</p>)}
+          </>
+        )}
       </div>
     </ColorPickerContext.Provider>
   );