+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# [v1.0.3] - 2022-11-29
+
+### Added
+
+- add omitted heading
+- change breakpoint from 600 to 850px
+
+# [v1.0.2] - 2022-11-29
+
+### Added
+
+- solving state
+
+# [v1.0.1] - 2022-11-29
+
+### Added
+
+- mobile responsive
+- descriptions of steps to solve
+
 # [v1.0.0] - 2022-11-28
 
 ### Added